+++ src/zkevm_specs/evm/execution/error_gas_uint_overflow.py
@@ -1,0 +1,5 @@
+from ..instruction import Instruction, Transition
+
+
+def gas_uint_overflow(instruction: Instruction):
+    opcode = instruction.opcode_lookup(True)
