+++ src/zkevm_specs/evm/execution/error_oog_call.py
@@ -72,10 +72,9 @@ def oog_call(instruction: Instruction):
         + memory_expansion_gas_cost
     )
 
-    # verify gas is insufficient or gas cost is overflow
+    # verify gas is insufficient
     gas_not_enough, _ = instruction.compare(instruction.curr.gas_left, gas_cost, N_BYTES_GAS)
-    gas_cost_overflow = instruction.is_u64_overflow(gas_cost)
-    instruction.constrain_equal(gas_not_enough + gas_cost_overflow, FQ(1))
+    instruction.constrain_equal(gas_not_enough, FQ(1))
 
     # current call must be failed.
     instruction.constrain_equal(
